Keyguard: LaterLock: improve the handling of policy timeout
If the security policy specifies a maximumTimeToLock, the
previous implementation would consider the screen's
displayTimeout as having elapsed, even if the user simply
pushed the power button to turn the screen off.

Consider the case of a 30 minute policy timeout and a 30 minute
display timeout.  In this case, pressing the power button will
cause the phone to be immediately locked (even if the user
requested it remain unlocked for 30 minutes).

This commit differentiates these two cases (display timeout
vs. user initiated display off) and only counts the displayTimeout
period if it wasn't user initiated.  After this change, the
lock would only occur 30 minutes after the power button was
pushed.

Note: if the user requests, in the Settings app, that the
Daydream start immediately, it will still consider the
displayTimeout to have occurred because we don't have enough
context at this level to tell whethr or not Daydream was
user initiated or due to a displayTimeout.

Change-Id: I8ea971c75685b2c2b9d4ec050a7bce6e28f45746